Optimize LoC
